requirements.txt
Django>=5.0
django-cors-headers>=4.3
PyJWT>=2.8
bcrypt>=4.1
Pillow>=10.0
PyMySQL>=1.1
cryptography>=41.0
cloudinary>=1.40.0
django-cloudinary-storage>=0.3.0
